ADC reading for thermistors
Replaced RAMPS 4.7k to 10k resistor

+++ Marlin/src/HAL/HAL_ESP32/HAL_ESP32.h
@@ -42,7 +42,7 @@
 // --------------------------------------------------------------------------
 
 /** result of last ADC conversion */
-extern int HAL_adc_result;
+extern uint16_t HAL_adc_result;
 
 // --------------------------------------------------------------------------
 // Public functions
